requirements.txt
psycopg2-binary
pillow
djangorestframework~=3.14.0
python-dotenv
django-filter~=23.2
Django
djangorestframework-simplejwt
coverage